requirements.txt
asn1crypto==0.24.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
defusedxml==0.5.0
Django==2.1
django-allauth==0.37.1
django-cors-headers==2.4.0
django-rest-auth==0.9.3
django-rest-knox==3.1.5
django-webpack-loader==0.6.0
django-widget-tweaks==1.4.3
djangorestframework==3.8.2
idna==2.7
oauthlib==2.1.0
pycparser==2.18
pyOpenSSL==18.0.0
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
urllib3==1.23
